refactor: location search query hook 분리
